requirements.txt
agno==1.1.9
ccxt==4.4.65
pandas==2.2.3
Pillow==11.1.0
python-dotenv==1.0.1
pytz==2025.1
Requests==2.32.3
streamlit==1.43.1
ta==0.11.0
websockets==14.2
